Proof-of-concept of synthetic input support.
Synthetic inputs are React components that handle input,
but are not literally <input> nodes at their root level.
A synthetic component may be rendered, for example, as a
<div> containing an <input> in some arbitrary sub-location.

This patch allows adapted classes to be configured with a
query selector that helps Reagent locate the correct input
node to update when setting values for the component as a
whole.